Docker/task2/services/web/requirements.txt
Flask==1.1.2
requests
psycopg2-binary
